requirements.txt
numpy
tensorflow-gpu
asyncio
requests